Move CqlLibraryString
